% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/addpseudodata.R
\name{addpseudodata}
\alias{addpseudodata}
\title{Adds pseudodata information into inputed data structures}
\usage{
addpseudodata(data = list(), units = list(), label = list(),
  weights = list())
}
\arguments{
\item{data}{structure with data}

\item{units}{structure with data units}

\item{label}{structure with data labels}

\item{weights}{structure with weights}
}
\value{
structures with data, units, labael and weights
}
\description{
Adds the pseudodata information and weights for purposes of the regression
}
\examples{
[data, units, label, weight] <- addpseudodata();
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{setweights}}
}

