% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/C2K.R
\name{C2K}
\alias{C2K}
\title{Conversion of Celsius to Kelvin}
\usage{
C2K(C)
}
\arguments{
\item{C}{numeric temperature in degrees Celsius}
}
\value{
temperature in Kelvin
}
\description{
Computes Kelvin from temperatures defined in Celsius
}
\examples{
C2K(20)
}
\seealso{
Other miscelaneous functions: \code{\link{K2C}};
  \code{\link{beta0}}; \code{\link{get_lb}};
  \code{\link{get_ue0}}; \code{\link{tempcorr}}
}

