% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dget_lbarb2.R
\name{dget_lbarb2}
\alias{dget_lbarb2}
\title{Computes derivative d delta/dx}
\usage{
dget_lbarb2(x, delta, pars)
}
\arguments{
\item{x}{scalar x = g/(g + e)}

\item{delta}{scalar delta = x e_H/ (1 - kap)g}

\item{pars}{data.frame with lbarb, xb, xb3 (xb^1/3), k}
}
\value{
scalar with derivative value d delta/ dx
}
\description{
Obtains the derivative d delta/dx from lbarb, xb and k.
}
\examples{
dget_lbarb2(10^(-6), 0, c(lbarb = 0.003, xb = 10/11, xb3 = (10/11)^(1/3), k = 1))
}
\seealso{
Other scaled get functions: \code{\link{fnget_lbarb2}};
  \code{\link{get_lbarb2}}; \code{\link{get_lbarb}};
  \code{\link{initial_scaled_reserve}}
}

