% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estim_options.R
\name{estim_options}
\alias{estim_options}
\title{Sets options for estim_pars}
\usage{
estim_options(key = "inexistent", val = "")
}
\arguments{
\item{key}{string with option to set}

\item{val}{value of the option}
}
\description{
Sets options for estimation one by one
}
\details{
no input: print values to screen

one input:

* "default": sets options at default values
* any other key (see below): print value to screen

two inputs:

* "filter": 1 - use filter (default); 0 - do not;
* "pars_init_method":
   0 - get initial estimates from automatized computation (default)
   1 - read initial estimates from .mat file (for continuation)
   2 - read initial estimates from pars_init file
* "pseudodata_pets":
   0 - put pseudodata together with data (default)
   1 - put it apart (only for multispecies estimation)
* "results_output":
   0 - prints results to screen (default)
   1 - prints results to screen, saves to .mat file
   2 - saves data to .mat file and graphs to .png files
   (prints results to screen using a customized results file when it exists)
 * "method": "nm" - use Nelder-Mead method; "no" - do not estimate;

 for other options see corresponding options file of the method (e.g. nmregr_options)
}
\examples{
estim_options("default")
}
\seealso{
Other regression functions: \code{\link{nmregr_options}}
}

