% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldnm_wtxt.R
\name{fieldnm_wtxt}
\alias{fieldnm_wtxt}
\title{Searches for fields with a given name in a multilevel structure}
\usage{
fieldnm_wtxt(data = list(), str = "")
}
\arguments{
\item{data}{data frame with fields}

\item{str}{string with field name}
}
\value{
list of strings with fields including the field name in str
}
\description{
Creates a list of field names of a data frame with str
}
\examples{
nmpsd <- fieldnm_wtxt(data, "psd")
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addpseudodata}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

