% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fieldnmnst.R
\name{fieldnmnst}
\alias{fieldnmnst}
\title{Creates a list of field names of a structure}
\usage{
fieldnmnst(st)
}
\arguments{
\item{st}{data frame with fields}
}
\value{
list of list of strings with fields including the field name in str
}
\description{
Creates a list of field names of a structure
}
\examples{
nst <- fieldnmnst(st)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{get_d_V}}; \code{\link{parscomp}};
  \code{\link{predict_pseudodata}};
  \code{\link{print_filterflag}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

