% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_std.R
\name{filter_std}
\alias{filter_std}
\title{Filters for allowed parameters of standard DEB model without acceleration}
\usage{
filter_std(par)
}
\arguments{
\item{par}{data frame with parameter values}
}
\value{
list with filter and flag
}
\description{
Checks if parameter values are in the allowable part of the parameter space of
standard DEB model without acceleration. Meant to be run in the estimation procedure
}
\details{
The flag is an indicator of reason for not passing the filter and it means
0: parameters pass the filter
1: some parameter is negative
2: some kappa is larger than 1
3: growth efficiency is larger than 1
4: maturity levels do not increase during life cycle
5: puberty cannot be reached
}
\examples{
filter_std(par)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{get_d_V}}; \code{\link{print_filterflag}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

