% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fnget_lambdab2.R
\name{fnget_lambdab2}
\alias{fnget_lambdab2}
\title{Computes f using the ode solver for delta(x), for finding lambdab}
\usage{
fnget_lambdab2(lambdab, pars)
}
\arguments{
\item{lambdab}{scalar with scaled length at birth (lambdab = lb/ g)}

\item{pars}{data.frame with lambdab, xb, xb3 (xb^1/3), k}
}
\value{
scalar with function f which when zero indicates lambdab
}
\description{
Computes f using the ode solver for delta(x), for finding lambdab.
}
\examples{
fnget_lambdab2(0.03, c(xb = 10/11, xb3 = (10/11)^(1/3), vvHb = 0.001, k = 1))
}
\seealso{
Other scaled get functions: \code{\link{dget_lambdab2}};
  \code{\link{dget_lbarb2}}; \code{\link{fnget_lbarb2}};
  \code{\link{get_lambdab2}}; \code{\link{get_lambdab}};
  \code{\link{get_lbarb2}}; \code{\link{get_lbarb}};
  \code{\link{initial_scaled_reserve}}
}

