% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_d_V.R
\name{get_d_V}
\alias{get_d_V}
\title{Sets chemical parameters and text for units and labels}
\usage{
get_d_V(phylum, class)
}
\arguments{
\item{phylum}{string with species phylum}

\item{class}{string with species class}
}
\value{
list with d_V and info
}
\description{
Sets chemical parameters and text for units and labels
}
\details{
Calls get_d_V to set specific density of structure. For a specific density of wet mass of 1 g/cm^3,
a specific density of d_E = d_V = 0.1 g/cm^3 means a dry-over-wet weight ratio of 0.1
}
\examples{
get_d_V("Chordata", "Mammalia")
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{parscomp}}; \code{\link{print_filterflag}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

