% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_tm_s.R
\name{get_tm_s}
\alias{get_tm_s}
\title{Obtains scaled mean age at death for short growth periods}
\usage{
get_tm_s(pars, f = 1, lb = NA, lp = NA)
}
\arguments{
\item{pars}{4 or 7-vector with parameters: [g lT ha sG] or [g k lT vHb vHp ha SG]}

\item{f}{optional scalar with scaled reserve density at birth (default f = 1)}

\item{lb}{optional scalar with scaled length at birth (default: lb is obtained from get_lb)}

\item{lp}{optional scalar with scaled length at puberty}
}
\value{
list with  scalar with scaled mean life span (tm),
scalar with survival probability at birth (if length p = 7) (Sb),
scalar with survival prabability at puberty (if length p = 7) (Sp)
and indicator equals 1 if successful convergence, 0 otherwise (info)
}
\description{
Obtains scaled mean age at death assuming a short growth period relative to the life span
}
\details{
Divide the result by the somatic maintenance rate coefficient to arrive at the mean age at death.
The variant get_tm_foetus does the same in case of foetal development.
If the input parameter vector has only 4 elements (for [g, lT, ha/ kM2, sG]),
it skips the calulation of the survival probability at birth and puberty.
}
\seealso{
Other scaled get functions: \code{\link{dget_lbarb2}};
  \code{\link{fnget_lbarb2}}; \code{\link{get_lbarb2}};
  \code{\link{get_lbarb}}; \code{\link{get_lb}};
  \code{\link{get_lp}}; \code{\link{get_tb}};
  \code{\link{get_tp}};
  \code{\link{initial_scaled_reserve}};
  \code{\link{reprod_rate}}
}

