% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ubare0.R
\name{get_ubarE0}
\alias{get_ubarE0}
\title{Computes initial scaled reserve density at birth}
\usage{
get_ubarE0(g = NA, k = NA, vbarHb = NA, eb = 1, lbarb = NA)
}
\arguments{
\item{eb:}{optional scalar with scaled reserbe density at birth}

\item{x1}{scalar with upper boundary for integration}
}
\value{
scalar with particular incomple beta function
}
\description{
Obtains the initial scaled reserve given the scaled reserve density at birth.
Function get_ue0 does so for eggs, get_ue0_foetus for foetuses.
Specification of length at birth as third input by-passes its computation,
so if you want to specify an initial value for this quantity, you should use get_lb directly.
}
\examples{
get_ubarE0(g = 10, lbarb = 0.01)
get_ubarE0(g = 10, k = 0.7, vbarHb = 5e-4)
}
\seealso{
Other get functions: \code{\link{get_ue0}}
}

