% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_ue0.R
\name{get_ue0}
\alias{get_ue0}
\title{Computes initial scaled reserve}
\usage{
get_ue0(pars, eb, lb0)
}
\arguments{
\item{pars}{1 or 3 -vector with parameters g, k_J/ k_M, v_H^b, see get_lb}

\item{x1}{scalar with upper boundary for integration}
}
\value{
scalar with particular incomple beta function
}
\description{
Obtains the initial scaled reserve given the scaled reserve density at birth.
Function get_ue0 does so for eggs, get_ue0_foetus for foetuses.
Specification of length at birth as third input by-passes its computation,
so if you want to specify an initial value for this quantity, you should use get_lb directly.
}
\examples{
beta0(0.1, 0.2)
}
\seealso{
Other get functions: \code{\link{get_ubarE0}}
}

