% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/initial_scaled_reserve.R
\name{initial_scaled_reserve}
\alias{initial_scaled_reserve}
\title{Gets initial scaled reserve}
\usage{
initial_scaled_reserve(f, p, Lb0)
}
\arguments{
\item{f}{n-vector with scaled functional responses}

\item{pars}{3-vector with parameters: g, k, vv_H^b (see below)}

\item{lb0}{optional n-vector with lengths at birth}
}
\value{
n-vector with initial scaled reserve: M_E^0/ {J_EAm} (U0), n-vector with length at birth 8Lb) and n-vector with 1's if successful, 0's otherwise (info)
}
\description{
Obtains scaled length at birth, given the scaled reserve density at birth.
}
\examples{
get_lambdab(c(g = 10, k = 1, vvHb = 0.0005), 1)
}
\seealso{
Other scaled get functions: \code{\link{dget_lambdab2}};
  \code{\link{dget_lbarb2}}; \code{\link{fnget_lambdab2}};
  \code{\link{fnget_lbarb2}}; \code{\link{get_lambdab2}};
  \code{\link{get_lambdab}}; \code{\link{get_lbarb2}};
  \code{\link{get_lbarb}}
}

