% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/K2C.R
\name{K2C}
\alias{K2C}
\title{Conversion of Kelvin to Celsius}
\usage{
K2C(K)
}
\arguments{
\item{K}{numeric temperature in degrees Kelvin}
}
\value{
temperature in Kelvin
}
\description{
Computes Celsius from temperatures given in Kelvin
}
\examples{
K2C(293.15)
}
\seealso{
Other miscelaneous.functions: \code{\link{C2K}};
  \code{\link{beta0}}; \code{\link{get_lambdab}};
  \code{\link{get_lb}}; \code{\link{get_ue0}};
  \code{\link{tempcorr}}
}

