% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mre_st.R
\name{mre_st}
\alias{mre_st}
\title{Computes mean relative error}
\usage{
mre_st(func, par, data, auxData, weights)
}
\arguments{
\item{func}{string with predict file name}

\item{par}{data frame with parameter values}

\item{data}{data frame with data values}

\item{auxData}{data frame with auxiliary data values}

\item{weights}{data frame with values of weights}
}
\description{
Computes relative errors and mean relative error for using data and predictions
}
\examples{
results_pets(par, metaPar, txtPar, data, auxData, metaData, txtData, weights)
}

