% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mydata_my_pet.R
\name{mydata_my_pet}
\alias{mydata_my_pet}
\title{Sets referenced data}
\usage{
mydata_my_pet()
}
\value{
list with data, auxData, metaData, txtData and weights
}
\description{
Sets data, pseudodata, metadata, auxdata, explanatory text, weights coefficients. Meant to be a template in add-my-pet
}
\examples{
mydata_my_pet()
}
\seealso{
Other add-my-pet template functions: \code{\link{pars_init_my_pet}}
}

