% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mydata_pets.R
\name{mydata_pets}
\alias{mydata_pets}
\title{Concatenates mydata files for several species}
\usage{
mydata_pets()
}
\value{
structure with data, auxData, metaData, txtData and weights for several pets
}
\description{
Concatenates mydata files for several species
}
\examples{
mydata_pets()
}
\seealso{
Other add-my-pet functions: \code{\link{estim_pars}}
}

