% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nmregr_options.R
\name{nmregr_options}
\alias{nmregr_options}
\title{Sets options for function nmregr}
\usage{
nmregr_options(key = "inexistent", val = "")
}
\arguments{
\item{key}{string with option to set}

\item{val}{value of the option}
}
\value{
1 if input is valid key, 0 if input is unknown key
}
\description{
Sets options for estimation one by one
}
\details{
no input: print values to screen

one input:

* "default": sets options at default values
* any other key (see below): print value to screen

two inputs:

* "report": 1 - to report steps to screen; 0 - not to;
* "max_step_number": maximum number of steps
* "max_fun_evals": maximum number of function evaluations
* "tol_simplex": tolerance for how close the simplex points must be together to call them the same
* "tol_tun": tolerance for how close the loss-function values must be together to call them the same
* "simplex_size": fraction added (subtracted if negative) to the free parameters when building the simplex
}
\examples{
nmregr_options("default")
}
\seealso{
Other regression functions: \code{\link{estim_options}}
}

