% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pars_init_my_pet.R
\name{pars_init_my_pet}
\alias{pars_init_my_pet}
\title{Sets (initial values for) parameters}
\usage{
pars_init_my_pet(metaData)
}
\arguments{
\item{metaData}{data frame with info about this entry (needed for names of phylum and class to get d_V)}
}
\value{
list with par (with values of parameters), metaPar (with information on metaparameters) and txtPar (with information on parameters)
}
\description{
Sets (initial values for) parameters$ Meant to be a template in add-my-pet
}
\examples{
pars_init_my_pet(metaData)
}
\seealso{
Other add-my-pet template functions: \code{\link{mydata_my_pet}};
  \code{\link{predict_my_pet}}
}

