% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parscomp.R
\name{parscomp}
\alias{parscomp}
\title{Computes compound parameters from primary parameters}
\usage{
parscomp(par)
}
\arguments{
\item{par}{data frame with parameter values}
}
\value{
list with compound parameters
}
\description{
Computes compound parameters from primary parameters that are frequently used
}
\examples{
parscomp(par)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{get_d_V}}; \code{\link{print_filterflag}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

