% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petregr_f.R
\name{petregr_f}
\alias{petregr_f}
\title{Calculates least squares estimates using Nelder Mead's simplex method using a filter}
\usage{
petregr_f(func, par, data, auxData, weights, filternm)
}
\arguments{
\item{func}{character string with name of user-defined function}

\item{par}{list with parameters}

\item{data}{list with data}

\item{auxData}{list with auxiliary data}

\item{weights}{list with weights}

\item{filternm}{character string with name of user-defined filter function}
}
\value{
list with list with parameters resulting from estimation procedure (par)
and indicator 1 if convergence has been successful or 0 otherwise (info)
}
\description{
Calculates least squares estimates using Nelder Mead's simplex method using a filter
}
\seealso{
Other add-my-pet functions: \code{\link{estim_pars}};
  \code{\link{mydata_pets}}; \code{\link{predict_pets}};
  \code{\link{printpar}}; \code{\link{printprd}};
  \code{\link{results_pets}}
}

