% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_my_pet.R
\name{predict_my_pet}
\alias{predict_my_pet}
\title{Obtains predictions, using parameters and data}
\usage{
predict_my_pet(par, data, auxData = list())
}
\arguments{
\item{par}{data frame with parameter values}

\item{data}{data frame with data values}

\item{auxData}{data frame with auxiliary data values}
}
\value{
list with prdData (data frame with values of predictions) and info (indicator for customized filters)
}
\description{
Obtains predictions, using parameters and data
}
\examples{
predict_my_pet(par, data, auxData)
}
\seealso{
Other add-my-pet template functions: \code{\link{mydata_my_pet}};
  \code{\link{pars_init_my_pet}}
}

