% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_pets.R
\name{predict_pets}
\alias{predict_pets}
\title{Concatenates predict files for several species}
\usage{
predict_pets(parGrp, data, auxData)
}
\arguments{
\item{parGrp}{data frame with parameter values of the group}

\item{data}{data frame with data values}

\item{auxData}{data frame with auxiliary data values}
}
\value{
structure with prdData and prdInfo for several pets
}
\description{
Concatenates predict files for several species
}
\seealso{
Other add-my-pet functions: \code{\link{estim_pars}};
  \code{\link{mydata_pets}}; \code{\link{petregr_f}};
  \code{\link{printpar}}; \code{\link{printprd}};
  \code{\link{results_pets}}
}

