% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_pseudodata.R
\name{predict_pseudodata}
\alias{predict_pseudodata}
\title{Predicts pseudodata values}
\usage{
predict_pseudodata(par, data, prdData)
}
\arguments{
\item{par}{data frame with parameter values}

\item{data}{data frame with data values}

\item{prdData}{data frame with prediction values}
}
\value{
structure with pseudodata predictions
}
\description{
Adds pseudodata predictions into predictions structure
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{fieldnmnst}}; \code{\link{get_d_V}};
  \code{\link{parscomp}}; \code{\link{print_filterflag}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

