% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_filterflag.R
\name{print_filterflag}
\alias{print_filterflag}
\title{Prints an explanation of the filter flag onto the screen}
\usage{
print_filterflag(flag)
}
\arguments{
\item{flag}{integer with code from filter}
}
\value{
list with updated par, units, label and free
}
\description{
Prints an explanation to the screen according to the flag produced by a filter.
Meant to be run in the estimation procedure for the seed parameter set
}
\examples{
pars_init_my_pet(metaData)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{filter_std}}; \code{\link{get_d_V}};
  \code{\link{rmpseudodata}}; \code{\link{setweights}}
}

