% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/results_pets.R
\name{results_pets}
\alias{results_pets}
\title{Prints results of estimation}
\usage{
results_pets(par, metaPar, txtPar, data, auxData, metaData, txtData, weights)
}
\arguments{
\item{par}{data frame with parameter values}

\item{metaPar}{data frame with metainformation on models}

\item{txtPar}{data frame with information on parameters}

\item{data}{data frame with data values}

\item{auxData}{data frame with auxiliary data values}

\item{metaData}{data frame with metainformation on the entry}

\item{txtData}{data frame with infromation on data}

\item{weights}{data frame with values of weights}
}
\description{
Prints the results of the esimation procedure in the screen, .mat file and makes figures of graphs
}
\examples{
results_pets(par, metaPar, txtPar, data, auxData, metaData, txtData, weights)
}
\seealso{
Other add-my-pet functions: \code{\link{estim_pars}};
  \code{\link{mydata_pets}}; \code{\link{petregr_f}}
}

