% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmpseudodata.R
\name{rmpseudodata}
\alias{rmpseudodata}
\title{Removes pseudodata information from inputed data structures}
\usage{
rmpseudodata(data = list())
}
\arguments{
\item{data}{structure with "psd" field to be removed}
}
\value{
structure with "psd" field removed
}
\description{
Removes pseudodata information from inputed data structures
}
\examples{
data <- rmpseudodata(data)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addchem}};
  \code{\link{addpseudodata}}; \code{\link{fieldnm_wtxt}};
  \code{\link{get_d_V}}; \code{\link{parscomp}};
  \code{\link{print_filterflag}}; \code{\link{setweights}}
}

