% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setweights.R
\name{setweights}
\alias{setweights}
\title{Sets automatically the weights for the data (to be used in a regression)}
\usage{
setweights(data, weights = list())
}
\arguments{
\item{data}{structure with data}

\item{weights}{structure with weights}
}
\value{
structure with weights
}
\description{
computes weights for given data and adds it to the weight structure
}
\details{
computes weights for given data and adds it to the weight structure
for the zero-variate data y the weight will be
\deqn{min(100, 1/ max(10^-6, y) ^2 \right)}{\min\left(100, \frac{1}{\max\left(10^{-6}, y\right ) ^2}}
for the uni-variate data y the weight will be
\deqn{1/ N \bar{y}^2}{\frac{1}{N \bar{y}^2}}
}
\examples{
setweights(data)
}
\seealso{
Other add-my-pet auxiliary functions: \code{\link{addpseudodata}}
}

