% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/beta0.R
\name{beta0}
\alias{beta0}
\title{Particular incomplete beta function}
\usage{
beta0(x0, x1)
}
\arguments{
\item{x0}{scalar with lower boundary for integration}

\item{x1}{scalar with upper boundary for integration}
}
\value{
scalar with particular incomple beta function
}
\description{
particular incomplete beta function:
}
\examples{
beta0(0.1, 0.2)
}
\seealso{
Other miscelaneous.functions: \code{\link{C2K}};
  \code{\link{K2C}}; \code{\link{get_lambdab}};
  \code{\link{get_lb}}; \code{\link{get_ue0}};
  \code{\link{tempcorr}}
}

