% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_lb.R
\name{get_lb}
\alias{get_lb}
\title{Computes initial scaled reserve}
\usage{
get_lb(p, eb, lb0 = NA)
}
\arguments{
\item{x0}{scalar with lower boundary for integration}

\item{x1}{scalar with upper boundary for integration}
}
\value{
scalar with particular incomple beta function
}
\description{
particular incomplete beta function:
}
\examples{
beta0(0.1, 0.2)
}
\seealso{
Other miscelaneous functions: \code{\link{C2K}};
  \code{\link{K2C}}; \code{\link{beta0}};
  \code{\link{get_ue0}}; \code{\link{tempcorr}}
}

