% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tempcorr.R
\name{tempcorr}
\alias{tempcorr}
\title{Conversion of Kelvin to Celsius}
\usage{
tempcorr(Temp, T_1, Tpars)
}
\arguments{
\item{T_1}{scalar with reference temperature}

\item{Tpars}{1-, 3- or 5-vector with temperature parameters}

\item{T}{vector with new temperatures}
}
\value{
vector with temperature correction factors that affect all rates
}
\description{
Calculates the factor with which physiological rates should be multiplied to go from a reference temperature to a given temperature
}
\examples{
tempcorr(c(330, 331, 332), 320, c(12000, 277, 318, 20000, 190000))
}

